WFP-1126 created table for storing workload calculations.

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsworkload/service/CapacityCalculator.kt
@@ -7,5 +7,5 @@ interface CapacityCalculator {
 
   fun calculate(totalPoints: BigInteger, availablePoints: BigInteger): BigDecimal
 
-  fun calculateAvailablePoints(availablePoints: BigDecimal, currentHours: BigDecimal, reductionHours: BigDecimal, defaultContractedHoursForGrade: BigDecimal): BigDecimal
+  fun calculateAvailablePoints(availablePoints: BigDecimal, currentHours: BigDecimal, reductionHours: BigDecimal, defaultContractedHoursForGrade: BigDecimal): BigInteger
 }
